# Conflicts:
#   Back-End/src/main/java/com/conseller/conseller/exception/GlobalExceptionHandler.java

+++ Back-End/src/main/java/com/conseller/conseller/auction/auction/dto/response/DetailAuctionResponse.java
@@ -20,7 +20,7 @@ public class DetailAuctionResponse {
 
     private String auctionUserProfileUrl;
 
-    private Integer auctionUserDeposit;
+    private Long auctionUserDeposit;
 
     private List<AuctionBidItemData> auctionBidList;
